:wrench: [WIP] Refactoring auth :wrench:
